requirements.txt
# Core Frameworks
streamlit>=1.30.0
gradio>=4.0.0

# Deep Learning (PyTorch Ecosystem)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Audio Processing
scipy>=1.11.0
soundfile>=0.12.0
librosa>=0.10.0

# Data & Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Development/Testing
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
wandb>=0.15.0
tensorboard>=2.13.0
